Added KS8000 & Clarification on websockets
I recently went through the process of setting up my Samsung KS8000 on HA.

THe instructions are actually wrong when advising to install websocket:

samsungtv component says to install websockets: pip3 install websocket-client <---this is not entirely correct

HA is installed in a virtual environment so all python packages are independent.
you have to activate the virtual environment.  I added some instructions to assist with that.